Exported joyride tour to separate class
